refactor: Character, Square 컴포넌트 내 로직 커스텀 훅으로 분리 (useTouchForMobile, useDragForWeb, useDropForWeb)
