+++ frontend/src/components/play/Square.tsx
@@ -1,10 +1,8 @@
-import { CharacterInfoInterface } from 'types/play';
-import 'styles/components/play/square.scss';
-import Character from 'components/play/Character';
+import { CharacterInfoInterface, CharacterPosition } from 'types/play';
 import { useTouchEndListener } from 'hooks/useTouchEndListener';
-import { moveCharacterForWeb } from 'utils/moveCharacterForWeb';
-import { usePlayStore } from 'stores/playStore';
-import { CharacterPosition } from 'types/play';
+import { useDropForWeb } from 'hooks/useDropForWeb';
+import Character from 'components/play/Character';
+import 'styles/components/play/square.scss';
 
 interface SquareProps {
   row: number;
@@ -18,23 +16,9 @@ interface SquareProps {
 }
 
 const Square = ({ row, col, characterInfo, updateBoard }: SquareProps) => {
-  const { prevPosition, setSelectedCharacterKey, setPrevPosition } =
-    usePlayStore();
-  const nextPosition = { row, col };
-
-  // 모바일 환경
   useTouchEndListener(row, col, updateBoard);
 
-  // 웹 환경
-  const handleDrop = (e: React.DragEvent<HTMLImageElement>) => {
-    moveCharacterForWeb(e, prevPosition, nextPosition, updateBoard);
-    setSelectedCharacterKey(null);
-    setPrevPosition({ row: null, col: null });
-  };
-
-  const handleDragOver = (e: React.DragEvent<HTMLImageElement>) => {
-    e.preventDefault();
-  };
+  const { handleDrop, handleDragOver } = useDropForWeb(row, col, updateBoard);
 
   return (
     <div
